split accesscontrol enumerability + enumerability tests
